add delete skill
Added a delete button to the all skills component,
with an onClick listener that will call
this.handleDelete, passing in this, and the id of
the skill to be deleted. handleDelete passes the 
data to the body via props.

the bodies' handleDelete method makes an ajax call
to remove the idea from the database, and upon
success, will call removeIdeaFromDOM, which will 
do just that.